struct Reviewer is now not a inner struct of Config
