+++ Sources/MonchCLI/Domain/ValueObject/Config.swift
@@ -53,24 +53,6 @@ extension Config {
     }
 }
 
-extension Config {
-    struct Reviewer: Decodable {
-        let name: String
-        let chatworkId: Int
-        let githubLogin: String
-
-        func isValid() -> Bool {
-            guard !name.isEmpty else {
-                fatalError("Reviewer の名前が空です。設定ファイルを確認してください。")
-            }
-            guard !githubLogin.isEmpty else {
-                fatalError("Reviewer の GitHub の名前が空です。設定ファイルを確認してください。")
-            }
-            return true
-        }
-    }
-}
-
 // MARK: - Validation Extension
 
 private extension String {
